Move Response type to submodule of jsonrpc
