+++ src/jsonrpc.rs
@@ -2,6 +2,7 @@
 
 pub use self::error::{Error, ErrorCode, Result};
 pub use self::request::{Request, RequestBuilder};
+pub use self::response::Response;
 pub(crate) use self::router::Router;
 pub use self::router::{FromParams, IntoResponse, Method};
 
@@ -12,10 +13,10 @@ use lsp_types::NumberOrString;
 use serde::de::{self, Deserializer};
 use serde::ser::Serializer;
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
 mod error;
 mod request;
+mod response;
 mod router;
 
 /// A unique ID used to correlate requests and responses together.
@@ -75,108 +76,6 @@ impl From<NumberOrString> for Id {
             NumberOrString::String(s) => Id::String(s),
         }
     }
-}
-
-/// A successful or failed JSON-RPC response.
-#[derive(Clone, PartialEq, Deserialize, Serialize)]
-pub struct Response {
-    jsonrpc: Version,
-    #[serde(flatten)]
-    kind: ResponseKind,
-    id: Id,
-}
-
-impl Response {
-    /// Creates a new successful response from a request ID and `Error` object.
-    pub const fn from_ok(id: Id, result: Value) -> Self {
-        Response {
-            jsonrpc: Version,
-            kind: ResponseKind::Ok { result },
-            id,
-        }
-    }
-
-    /// Creates a new error response from a request ID and `Error` object.
-    pub const fn from_error(id: Id, error: Error) -> Self {
-        Response {
-            jsonrpc: Version,
-            kind: ResponseKind::Err { error },
-            id,
-        }
-    }
-
-    /// Creates a new response from a request ID and either an `Ok(Value)` or `Err(Error)` body.
-    pub fn from_parts(id: Id, body: Result<Value>) -> Self {
-        match body {
-            Ok(result) => Response::from_ok(id, result),
-            Err(error) => Response::from_error(id, error),
-        }
-    }
-
-    /// Splits the response into a request ID paired with either an `Ok(Value)` or `Err(Error)` to
-    /// signify whether the response is a success or failure.
-    pub fn into_parts(self) -> (Id, Result<Value>) {
-        match self.kind {
-            ResponseKind::Ok { result } => (self.id, Ok(result)),
-            ResponseKind::Err { error } => (self.id, Err(error)),
-        }
-    }
-
-    /// Returns `true` if the response indicates success.
-    pub const fn is_ok(&self) -> bool {
-        matches!(self.kind, ResponseKind::Ok { .. })
-    }
-
-    /// Returns `true` if the response indicates failure.
-    pub const fn is_error(&self) -> bool {
-        !self.is_ok()
-    }
-
-    /// Returns the `result` value, if it exists.
-    ///
-    /// This member only exists if the response indicates success.
-    pub const fn result(&self) -> Option<&Value> {
-        match &self.kind {
-            ResponseKind::Ok { result } => Some(result),
-            _ => None,
-        }
-    }
-
-    /// Returns the `error` value, if it exists.
-    ///
-    /// This member only exists if the response indicates failure.
-    pub const fn error(&self) -> Option<&Error> {
-        match &self.kind {
-            ResponseKind::Err { error } => Some(error),
-            _ => None,
-        }
-    }
-
-    /// Returns the corresponding request ID, if known.
-    pub const fn id(&self) -> &Id {
-        &self.id
-    }
-}
-
-impl Debug for Response {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        let mut d = f.debug_struct("Response");
-        d.field("jsonrpc", &self.jsonrpc);
-
-        match &self.kind {
-            ResponseKind::Ok { result } => d.field("result", result),
-            ResponseKind::Err { error } => d.field("error", error),
-        };
-
-        d.field("id", &self.id).finish()
-    }
-}
-
-#[derive(Clone, PartialEq, Deserialize, Serialize)]
-#[serde(untagged)]
-enum ResponseKind {
-    Ok { result: Value },
-    Err { error: Error },
 }
 
 /// An incoming or outgoing JSON-RPC message.
